Create class PDFDocument and pass function from PDFCore to PDFDocument
